fix: remove redundant cast and bound d0 fallback tests to 10 s
- stats_service: drop `as u64` on num_threads() (already u64 — clippy warning)
- interop d0 tests: wrap client.dial() in timeout(10 s) for the two
  fallback tests; do_fallback uses CONNECTION_IDLE_TIMEOUT (300 s), so
  without a bound the tests took ~300 s each.  Dropping the dial() future
  closes the client TCP stream → relay's copy_bidirectional_with_idle
  detects client EOF and returns, so accept_direct resolves in ≤ 10 s.
- Remove unused dev_self_signed / tls_accept imports from interop.rs
- cargo fmt: collapse Hysteria2UdpSession::send signature to one line

+++ crates/blackwire-api/src/stats_service.rs
@@ -65,7 +65,7 @@ impl StatsService for StatsServiceImpl {
             sys: rss,
             mallocs: 0,
             frees: 0,
-            live_objects: runtime_stats::num_threads() as u64,
+            live_objects: runtime_stats::num_threads(),
             pause_total_ns: 0,
             uptime: runtime_stats::uptime_secs(),
         }))
